Alteração do header, login e footer.
