+++ src/Components/Footer.module.css
@@ -1,7 +1,7 @@
 .footer {
   background: #fb1;
   padding: 3rem 1rem 0 1rem;
-  height: 10rem;
+  height: 5rem;
   text-align: center;
   color: #764701;
 }
