--- src/pages/Menu/index.css
@@ -1,50 +1,0 @@
-.hamburgers-thumb {
-    width: 100px;
-    height: 60px;
-    
-}
-
-.hamburgers-thumb img {
-    width: 70%;
-    height: 70%;
-}
-
-.container {
-    perspective: 400px;
-    width: 20%;
-    height: 190px;
-    margin: 0 2.5% 40px 2.5%;
-}
-
-.card:nth-child(3n) {
-    margin-right: 0;
-}
-
-.card:nth-child(3n+1) {
-    margin-left: 0;
-}
-
-
-.card-container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    -webkit-backface-visibility: hidden;
-    backface-visibility: hidden;
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: white;
-    text-align: center;
-    /*box-shadow: 4px 4px 20px rgba(197, 143, 43, 0.919);*/
-}
-
-
-
-.card-container p {
-    margin: 5px 0;
-    font-size: 14px;
-}
-
